requirements.txt
# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# Deep Learning
tensorflow>=2.7.0
keras>=2.7.0
torch>=1.10.0
transformers>=4.12.0

# API
python-binance>=1.0.15
alpha_vantage>=2.3.1
ta>=0.7.0
pandas-ta>=0.3.14b0
ccxt>=1.60.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
websockets>=9.1

# Visualization
plotly>=5.3.1
dash>=2.0.0
streamlit>=1.2.0
python-telegram-bot>=13.7

# Testing
pytest>=6.2.5

# Development
black>=21.9b0
flake8>=3.9.2
isort>=5.9.3
mkdocs>=1.2.3
mkdocs-material>=8.0.0
joblib>=1.1.0
tqdm>=4.62.0 